Move extender implementor down and use TreeItem
