+++ src/api/ZoweExplorerApiRegister.ts
@@ -12,7 +12,7 @@
 import { IProfileLoaded } from "@zowe/imperative";
 import { ZoweExplorerApi } from "./ZoweExplorerApi";
 import { ZosmfUssApi as ZosmfUssApi, ZosmfMvsApi, ZosmfJesApi } from "./ZoweExplorerZosmfApi";
-import { ZoweExplorerExtender } from "./ZoweExplorerExtender";
+import { ZoweExplorerExtender } from "../ZoweExplorerExtender";
 
 import * as nls from "vscode-nls";
 const localize = nls.config({messageFormat: nls.MessageFormat.file})();
